Refactored dataset details
