+++ src/app/datasets/dataset-detail/dataset-detail.component.ts
@@ -1,37 +1,13 @@
-import { ActivatedRoute } from "@angular/router";
-import { Component, Inject, OnDestroy, OnInit } from "@angular/core";
 import {
-  DatablocksAction,
-  DeleteAttachment,
-  UpdateAttachmentCaptionAction,
-  ClearFacetsAction,
-  AddAttachment
-} from "state-management/actions/datasets.actions";
-import { Job, User, Attachment, Dataset } from "shared/sdk/models";
-import { select, Store } from "@ngrx/store";
-import { AddKeywordFilterAction } from "state-management/actions/datasets.actions";
-import { SubmitAction } from "state-management/actions/jobs.actions";
-import { ShowMessageAction } from "state-management/actions/user.actions";
-import {
-  getError,
-  submitJob
-} from "../../state-management/selectors/jobs.selectors";
-import { Subscription, Observable } from "rxjs";
-import { Message, MessageType } from "state-management/models";
-import { getIsAdmin } from "state-management/selectors/users.selectors";
+  Component,
+  Inject,
+  OnInit,
+  Input,
+  Output,
+  EventEmitter
+} from "@angular/core";
+import { Attachment, Dataset } from "shared/sdk/models";
 import { APP_CONFIG, AppConfig } from "app-config.module";
-import { pluck, take } from "rxjs/operators";
-import { Router } from "@angular/router";
-import {
-  getCurrentAttachments,
-  getCurrentDatablocks,
-  getCurrentDataset,
-  getCurrentOrigDatablocks,
-  getCurrentDatasetWithoutOrigData,
-  getViewPublicMode
-} from "state-management/selectors/datasets.selectors";
-import { ReadFile } from "ngx-file-helpers";
-import { UserApi } from "shared/sdk";
 
 /**
  * Component to show details for a data set, using the
@@ -45,175 +21,28 @@ import { UserApi } from "shared/sdk";
   templateUrl: "./dataset-detail.component.html",
   styleUrls: ["./dataset-detail.component.scss"]
 })
-export class DatasetDetailComponent implements OnInit, OnDestroy {
-  datasetPid: string;
-  dataset$ = this.store.pipe(select(getCurrentDataset));
-  datasetwithout$ = this.store.pipe(select(getCurrentDatasetWithoutOrigData));
-  jwt$: Observable<any>;
+export class DatasetDetailComponent implements OnInit {
+  @Input() dataset: Dataset;
+  @Input() datasetWithout: any;
+  @Input() attachments: Attachment[];
 
-  private subscriptions: Subscription[] = [];
-  private routeSubscription = this.route.params
-    .pipe(pluck("id"))
-    .subscribe((id: string) => {
-      this.datasetPid = id;
-    });
-  public origDatablocks$ = this.store.pipe(select(getCurrentOrigDatablocks));
-  public datablocks$ = this.store.pipe(select(getCurrentDatablocks));
-  public attachments$ = this.store.pipe(select(getCurrentAttachments));
-  public isAdmin$ = this.store.pipe(select(getIsAdmin));
-  public viewPublic: boolean;
+  @Output() clickKeyword = new EventEmitter<string>();
+  @Output() clickProposal = new EventEmitter<string>();
+  @Output() clickSample = new EventEmitter<string>();
 
-  dataset: Dataset;
-  pickedFile: ReadFile;
-  attachment: Attachment;
-
-  constructor(
-    private router: Router,
-    private route: ActivatedRoute,
-    private store: Store<any>,
-    private userApi: UserApi,
-    @Inject(APP_CONFIG) public appConfig: AppConfig
-  ) {}
-
-  ngOnInit() {
-    const msg = new Message();
-
-    this.subscriptions.push(
-      this.store.pipe(select(submitJob)).subscribe(
-        ret => {
-          if (ret && Array.isArray(ret)) {
-            console.log(ret);
-          }
-        },
-        error => {
-          console.log(error);
-          msg.type = MessageType.Error;
-          msg.content = "Job not Submitted";
-          this.store.dispatch(new ShowMessageAction(msg));
-        }
-      )
-    );
-
-    this.subscriptions.push(
-      this.store.pipe(select(getError)).subscribe(err => {
-        if (err) {
-          msg.type = MessageType.Error;
-          msg.content = err.message;
-          this.store.dispatch(new ShowMessageAction(msg));
-        }
-      })
-    );
-
-    this.subscriptions.push(
-      this.store.pipe(select(getViewPublicMode)).subscribe(viewPublic => {
-        this.viewPublic = viewPublic;
-      })
-    );
-
-    this.subscriptions.push(
-      this.dataset$.subscribe(dataset => {
-        this.dataset = dataset;
-      })
-    );
-
-    if (this.viewPublic) {
-      this.store.dispatch(
-        new DatablocksAction(this.datasetPid, { isPublished: this.viewPublic })
-      );
-    } else {
-      this.store.dispatch(new DatablocksAction(this.datasetPid));
-    }
-
-    this.jwt$ = this.userApi.jwt();
+  onClickKeyword(keyword: string): void {
+    this.clickKeyword.emit(keyword);
   }
 
-  ngOnDestroy() {
-    this.routeSubscription.unsubscribe();
-    for (let i = 0; i < this.subscriptions.length; i++) {
-      this.subscriptions[i].unsubscribe();
-    }
-  }
-
-  resetDataset(dataset) {
-    if (!confirm("Reset datablocks?")) {
-      return null;
-    }
-    this.store
-      .pipe(
-        select(state => state.root.user.currentUser),
-        take(1)
-      )
-      .subscribe((user: User) => {
-        const job = new Job();
-        job.emailJobInitiator = user.email;
-        job.jobParams = {};
-        job.jobParams["username"] = user.username;
-        job.creationTime = new Date();
-        job.type = "reset";
-        const fileObj = {};
-        const fileList = [];
-        fileObj["pid"] = dataset["pid"];
-        if (dataset["datablocks"]) {
-          dataset["datablocks"].map(d => {
-            fileList.push(d["archiveId"]);
-          });
-        }
-        fileObj["files"] = fileList;
-        job.datasetList = [fileObj];
-        console.log(job);
-        this.store.dispatch(new SubmitAction(job));
-      });
-  }
-
-  updateCaption(datasetId: string, attachmentId: string, caption: string) {
-    this.store.dispatch(
-      new UpdateAttachmentCaptionAction(datasetId, attachmentId, caption)
-    );
-  }
-
-  delete(dataset_id, dataset_attachment_id) {
-    this.store.dispatch(
-      new DeleteAttachment(dataset_id, dataset_attachment_id)
-    );
-  }
-
-  onClickProp(proposalId: string): void {
-    const id = encodeURIComponent(proposalId);
-    this.router.navigateByUrl("/proposals/" + id);
+  onClickProposal(proposalId: string): void {
+    this.clickProposal.emit(proposalId);
   }
 
   onClickSample(sampleId: string): void {
-    const id = encodeURIComponent(sampleId);
-    this.router.navigateByUrl("/samples/" + id);
+    this.clickSample.emit(sampleId);
   }
 
-  onClickKeyword(keyword: string): void {
-    this.store.dispatch(new ClearFacetsAction());
-    this.store.dispatch(new AddKeywordFilterAction(keyword));
-    this.router.navigateByUrl("/datasets");
-  }
+  constructor(@Inject(APP_CONFIG) public appConfig: AppConfig) {}
 
-  onFileUploaderFilePicked(file: ReadFile) {
-    this.pickedFile = file;
-  }
-
-  onFileUploaderReadEnd(fileCount: number) {
-    if (fileCount > 0) {
-      this.attachment = {
-        thumbnail: this.pickedFile.content,
-        caption: this.pickedFile.name,
-        creationTime: new Date(),
-        id: null,
-        dataset: this.dataset,
-        datasetId: this.dataset.pid,
-        rawDatasetId: null,
-        derivedDatasetId: null,
-        proposal: null,
-        proposalId: null,
-        sample: null,
-        sampleId: null
-      };
-      this.store.dispatch(new AddAttachment(this.attachment));
-    }
-  }
+  ngOnInit() {}
 }
