+++ src/app/datasets/dataset-details-dashboard/dataset-details-dashboard.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy, Inject } from "@angular/core";
 import { Store, select } from "@ngrx/store";
-import { Dataset, UserApi } from "shared/sdk";
+import { Dataset, UserApi, User, Job, Attachment } from "shared/sdk";
 import {
   getCurrentDataset,
   getCurrentDatasetWithoutOrigData,
@@ -9,15 +9,27 @@ import {
   getCurrentAttachments,
   getViewPublicMode
 } from "state-management/selectors/datasets.selectors";
-import { getIsAdmin } from "state-management/selectors/users.selectors";
+import {
+  getIsAdmin,
+  getCurrentUser
+} from "state-management/selectors/users.selectors";
 import { ActivatedRoute, Router } from "@angular/router";
 import { Subscription, Observable } from "rxjs";
-import { pluck } from "rxjs/operators";
+import { pluck, take } from "rxjs/operators";
 import { APP_CONFIG, AppConfig } from "app-config.module";
 import { Message, MessageType } from "state-management/models";
 import { submitJob, getError } from "state-management/selectors/jobs.selectors";
 import { ShowMessageAction } from "state-management/actions/user.actions";
-import { DatablocksAction } from "state-management/actions/datasets.actions";
+import {
+  DatablocksAction,
+  UpdateAttachmentCaptionAction,
+  DeleteAttachment,
+  AddAttachment,
+  ClearFacetsAction,
+  AddKeywordFilterAction
+} from "state-management/actions/datasets.actions";
+import { SubmitAction } from "state-management/actions/jobs.actions";
+import { ReadFile } from "ngx-file-helpers";
 
 @Component({
   selector: "dataset-details-dashboard",
@@ -33,10 +45,94 @@ export class DatasetDetailsDashboardComponent implements OnInit, OnDestroy {
   jwt$: Observable<any>;
 
   dataset: Dataset;
-  datasetPid: string;
   viewPublic: boolean;
+  pickedFile: ReadFile;
+  attachment: Attachment;
 
   private subscriptions: Subscription[] = [];
+
+  onClickKeyword(keyword: string) {
+    this.store.dispatch(new ClearFacetsAction());
+    this.store.dispatch(new AddKeywordFilterAction(keyword));
+    this.router.navigateByUrl("/datasets");
+  }
+
+  onClickProposal(proposalId: string) {
+    const id = encodeURIComponent(proposalId);
+    this.router.navigateByUrl("/proposals/" + id);
+  }
+
+  onClickSample(sampleId: string) {
+    const id = encodeURIComponent(sampleId);
+    this.router.navigateByUrl("/samples/" + id);
+  }
+
+  resetDataset(dataset: Dataset) {
+    if (!confirm("Reset datablocks?")) {
+      return null;
+    }
+    this.store
+      .pipe(
+        select(getCurrentUser),
+        take(1)
+      )
+      .subscribe((user: User) => {
+        const job = new Job();
+        job.emailJobInitiator = user.email;
+        job.jobParams = {};
+        job.jobParams["username"] = user.username;
+        job.creationTime = new Date();
+        job.type = "reset";
+        const fileObj = {};
+        const fileList = [];
+        fileObj["pid"] = dataset["pid"];
+        if (dataset["datablocks"]) {
+          dataset["datablocks"].map(d => {
+            fileList.push(d["archiveId"]);
+          });
+        }
+        fileObj["files"] = fileList;
+        job.datasetList = [fileObj];
+        console.log(job);
+        this.store.dispatch(new SubmitAction(job));
+      });
+  }
+
+  onFileUploaderFilePicked(file: ReadFile) {
+    this.pickedFile = file;
+  }
+
+  onFileUploaderReadEnd(fileCount: number) {
+    if (fileCount > 0) {
+      this.attachment = {
+        thumbnail: this.pickedFile.content,
+        caption: this.pickedFile.name,
+        creationTime: new Date(),
+        id: null,
+        dataset: this.dataset,
+        datasetId: this.dataset.pid,
+        rawDatasetId: null,
+        derivedDatasetId: null,
+        proposal: null,
+        proposalId: null,
+        sample: null,
+        sampleId: null
+      };
+      this.store.dispatch(new AddAttachment(this.attachment));
+    }
+  }
+
+  updateCaption(datasetId: string, attachmentId: string, caption: string) {
+    this.store.dispatch(
+      new UpdateAttachmentCaptionAction(datasetId, attachmentId, caption)
+    );
+  }
+
+  deleteAttachment(dataset_id, dataset_attachment_id) {
+    this.store.dispatch(
+      new DeleteAttachment(dataset_id, dataset_attachment_id)
+    );
+  }
 
   constructor(
     @Inject(APP_CONFIG) public appConfig: AppConfig,
@@ -51,7 +147,15 @@ export class DatasetDetailsDashboardComponent implements OnInit, OnDestroy {
 
     this.subscriptions.push(
       this.route.params.pipe(pluck("id")).subscribe((id: string) => {
-        this.datasetPid = id;
+        if (id) {
+          if (this.viewPublic) {
+            this.store.dispatch(
+              new DatablocksAction(id, { isPublished: this.viewPublic })
+            );
+          } else {
+            this.store.dispatch(new DatablocksAction(id));
+          }
+        }
       })
     );
 
@@ -92,14 +196,6 @@ export class DatasetDetailsDashboardComponent implements OnInit, OnDestroy {
         this.viewPublic = viewPublic;
       })
     );
-
-    if (this.viewPublic) {
-      this.store.dispatch(
-        new DatablocksAction(this.datasetPid, { isPublished: this.viewPublic })
-      );
-    } else {
-      this.store.dispatch(new DatablocksAction(this.datasetPid));
-    }
 
     this.jwt$ = this.userApi.jwt();
   }
